Resolve runtime errors

+++ sw/model/0-1-golden-alg/core.cpp
@@ -12,7 +12,7 @@ core::core(sc_module_name name) : sc_module(name) {
 /** Process the first five bytes of each array argument. */
 uint32_t core::calculate_row_result(uint32_t carry, uint8_t *kern_row, uint8_t kern_dim, uint8_t *group) {
     for (int i = 0; i < kern_dim; ++i) {
-        carry += kern_row[i] * group[i];
+        carry += (uint32_t)kern_row[i] * (uint32_t)group[i];
     }
     
     return carry;
